Extract the LoginForm from loginpage module
Also convert LoginPage and LoginForm to use styled-components.
