+++ gsa/src/web/pages/login/loginpage.js
@@ -28,28 +28,20 @@ import {connect} from 'react-redux';
 
 import {withRouter} from 'react-router-dom';
 
-import glamorous from 'glamorous';
+import styled from 'styled-components';
 
 import _ from 'gmp/locale';
 
 import logger from 'gmp/log';
 
-import {KeyCode} from 'gmp/utils/event';
-import {isDefined} from 'gmp/utils/identity';
 import {isEmpty} from 'gmp/utils/string';
 
 import compose from 'web/utils/compose';
 import PropTypes from 'web/utils/proptypes';
-import withGmp from 'web/utils/withGmp';
 import Theme from 'web/utils/theme';
-
-import FormGroup from 'web/components/form/formgroup';
-import PasswordField from 'web/components/form/passwordfield';
-import SubmitButton from 'web/components/form/submitbutton';
-import TextField from 'web/components/form/textfield';
+import withGmp from 'web/utils/withGmp';
 
 import GBIcon from 'web/components/icon/greenboneicon';
-import Icon from 'web/components/icon/icon';
 
 import Layout from 'web/components/layout/layout';
 
@@ -58,193 +50,49 @@ import Header from 'web/components/structure/header';
 
 import {updateTimezone} from 'web/store/usersettings/actions';
 
+import LoginForm from './loginform';
+
 const log = logger.getLogger('web.login');
 
-const panelcss = {
-  marginTop: '5px',
-  marginBottom: '5px',
-  paddingBottom: '10px',
-  fontSize: '9pt',
-};
+const GreenboneIcon = styled(GBIcon)`
+  width: 30vh;
+  margin: 30px auto;
+  position: sticky;
+`;
 
-const Panel = glamorous.div(
-  'login-panel',
-  panelcss,
-);
+const LoginLayout = styled(Layout)`
+  height: 100%;
+  width: 420px;
+  margin: 0 auto;
+  padding: 20px 20px 0px 20px;
+`;
 
-const LoginPanel = glamorous(Layout)(
-  'login-panel',
-  panelcss,
-);
+const StyledLayout = styled(Layout)`
+  flex-direction: column;
+  height: 100vh;
+`;
 
-const Div = glamorous.div({
-  display: 'flex',
-  flexDirection: 'row',
-  margin: '0 auto',
-});
+const LoginHeader = styled(Header)`
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+`;
 
-const Error = glamorous.p(
-  'error',
-  {
-    color: Theme.warningRed,
-    fontWeight: 'bold',
-    textAlign: 'center',
-    margin: '10px',
-  }
-);
+const MenuSpacer = styled.div`
+  background: ${Theme.darkGray};
+  position: absolute;
+  top: 42px;
+  left: 0;
+  right: 0;
+  height: 35px;
+  zIndex: ${Theme.Layers.menu};
+`;
 
-class LoginForm extends React.Component {
-
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      username: '',
-      password: '',
-    };
-
-    this.handleSubmit = this.handleSubmit.bind(this);
-    this.handleValueChange = this.handleValueChange.bind(this);
-    this.handleKeyDown = this.handleKeyDown.bind(this);
-  }
-
-  handleSubmit() {
-    const {onSubmit} = this.props;
-
-    if (!isDefined(onSubmit)) {
-      return;
-    }
-
-    const {username, password} = this.state;
-    onSubmit(username, password);
-  }
-
-  handleValueChange(value, name) {
-    this.setState({[name]: value});
-  }
-
-  handleKeyDown(event) {
-    if (event.keyCode === KeyCode.ENTER) {
-      this.handleSubmit(event);
-    }
-  }
-
-  render() {
-    const {error} = this.props;
-    const {username, password} = this.state;
-    const protocol_insecure = window.location.protocol !== 'https:';
-    return (
-      <React.Fragment>
-        {protocol_insecure &&
-          <Panel>
-            <Error>{_('Warning: Connection unencrypted')}</Error>
-            <p>{_('The connection to this GSA is not encrypted, allowing ' +
-              'anyone listening to the traffic to steal your credentials.')}</p>
-            <p>{_('Please configure a TLS certificate for the HTTPS service ' +
-              'or ask your administrator to do so as soon as possible.')}</p>
-          </Panel>
-        }
-
-        <LoginPanel
-          flex="column"
-          align="space-around"
-        >
-          <Layout flex="row">
-            <Div>
-              <StyledIcon img="login-label.png" size="default"/>
-              <Layout flex="column">
-                <FormGroup title={_('Username')} titleSize="4">
-                  <TextField
-                    name="username"
-                    grow="1"
-                    placeholder={_('e.g. johndoe')}
-                    value={username}
-                    autoFocus="autofocus"
-                    tabIndex="1"
-                    onChange={this.handleValueChange}
-                  />
-                </FormGroup>
-                <FormGroup title={_('Password')} titleSize="4">
-                  <PasswordField
-                    name="password"
-                    grow="1"
-                    placeholder={_('Password')}
-                    value={password}
-                    onKeyDown={this.handleKeyDown}
-                    onChange={this.handleValueChange}
-                  />
-                </FormGroup>
-                <FormGroup size="4" offset="4">
-                  <SubmitButton
-                    flex
-                    grow
-                    title={_('Login')}
-                    onClick={this.handleSubmit}
-                  />
-                </FormGroup>
-              </Layout>
-            </Div>
-          </Layout>
-        </LoginPanel>
-        {isDefined(error) &&
-          <Panel>
-            <Error>{error}</Error>
-          </Panel>
-        }
-      </React.Fragment>
-    );
-  }
-}
-
-LoginForm.propTypes = {
-  error: PropTypes.string,
-  onSubmit: PropTypes.func,
-};
-
-const GreenboneIcon = glamorous(GBIcon)({
-  width: '30vh',
-  margin: '30px auto',
-  position: 'sticky',
-});
-
-const LoginLayout = glamorous(Layout)({
-  height: '100%',
-  width: '420px',
-  margin: '0 auto',
-  padding: '20px 20px 0px 20px',
-});
-
-const StyledLayout = glamorous(Layout)({
-  flexDirection: 'column',
-  height: '100vh',
-});
-
-const LoginHeader = glamorous(Header)({
-  position: 'absolute',
-  top: 0,
-  left: 0,
-  right: 0,
-});
-
-const StyledIcon = glamorous(Icon)({
-  height: '95px',
-  marginLeft: '5px',
-});
-
-const MenuSpacer = glamorous.div({
-  background: '#393637',
-  position: 'absolute',
-  top: '42px',
-  left: 0,
-  right: 0,
-  height: '35px',
-  zIndex: '500',
-});
-
-const Wrapper = glamorous.div({
-    border: '1px solid #ddd',
-    padding: '10px',
-});
+const Wrapper = styled.div`
+  border: 1px solid #ddd;
+  padding: 10px;
+`;
 
 class LoginPage extends React.Component {
